Created taskverification/service.go for Service related code
